moved the activities array out to App and then passed it into ActivitiesList as a prop called activities and other minor fixes
