+++ src/main.tsx
@@ -1,6 +1,3 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-
 import "./global.css";
 
 import { StrictMode } from "react";
